+++ app/lib/providers/developer_mode_provider.dart
@@ -6,6 +6,7 @@ import 'package:friend_private/providers/base_provider.dart';
 import 'package:friend_private/utils/alerts/app_snackbar.dart';
 import 'package:friend_private/utils/analytics/mixpanel.dart';
 import 'package:friend_private/utils/logger.dart';
+import 'package:friend_private/utils/other/validators.dart';
 
 class DeveloperModeProvider extends BaseProvider {
   final TextEditingController gcpCredentialsController = TextEditingController();
@@ -135,22 +136,6 @@ class DeveloperModeProvider extends BaseProvider {
     // getUserWebhookUrl(type: 'audio_bytes_websocket').then((url) => webhookWsAudioBytes.text = url);
     setIsLoading(false);
     notifyListeners();
-  }
-
-  bool isValidUrl(String url) {
-    const urlPattern = r'^(https?:\/\/)?([a-zA-Z0-9.-]+(:[a-zA-Z0-9.&%$-]+)*@)?'
-        r'((25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}'
-        r'(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)|'
-        r'([a-zA-Z0-9-]+\.)+[a-zA-Z]{2,63}(:[0-9]+)?(\/.*)?$';
-    return RegExp(urlPattern).hasMatch(url);
-  }
-
-  bool isValidWebSocketUrl(String url) {
-    const webSocketPattern = r'^(wss?:\/\/)?([a-zA-Z0-9.-]+(:[a-zA-Z0-9.&%$-]+)*@)?'
-        r'((25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}'
-        r'(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)|'
-        r'([a-zA-Z0-9-]+\.)+[a-zA-Z]{2,63}(:[0-9]+)?(\/.*)?$';
-    return RegExp(webSocketPattern).hasMatch(url);
   }
 
   void saveSettings() async {
